Add Sstp class
